hugolib: Remove siteWriter
