+++ hugolib/handler_file.go
@@ -39,7 +39,10 @@ type defaultHandler struct{ basicFileHandler }
 
 func (h defaultHandler) Extensions() []string { return []string{"*"} }
 func (h defaultHandler) FileConvert(f *source.File, s *Site) HandledResult {
-	s.w.writeDestFile(f.Path(), f.Contents)
+	err := s.publish(f.Path(), f.Contents)
+	if err != nil {
+		return HandledResult{err: err}
+	}
 	return HandledResult{file: f}
 }
 
@@ -48,6 +51,9 @@ type cssHandler struct{ basicFileHandler }
 func (h cssHandler) Extensions() []string { return []string{"css"} }
 func (h cssHandler) FileConvert(f *source.File, s *Site) HandledResult {
 	x := cssmin.Minify(f.Bytes())
-	s.w.writeDestFile(f.Path(), bytes.NewReader(x))
+	err := s.publish(f.Path(), bytes.NewReader(x))
+	if err != nil {
+		return HandledResult{err: err}
+	}
 	return HandledResult{file: f}
 }
